pyproject.toml
[project]
name = "rust-warp"
version = "0.1.0"
description = "High-performance raster reprojection engine in Rust with xarray/dask integration"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Jake Wilkins", email = "jake@epoch.blue" }
]
requires-python = ">=3.12"
keywords = ["geospatial", "reprojection", "raster", "xarray", "dask", "rust"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: GIS",
    "Programming Language :: Rust",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy>=1.24",
]

[project.optional-dependencies]
xarray = [
    "xarray>=2023.1",
    "dask[array]>=2023.1",
    "pyproj>=3.4",
]
all = [
    "rust-warp[xarray]",
    "odc-geo>=0.4",
]
dev = [
    "rust-warp[all]",
    "pytest>=7.0",
    "pytest-benchmark>=4.0",
    "rioxarray>=0.15",
    "rasterio>=1.3",
    "odc-geo>=0.4",
    "pyproj>=3.4",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "obstore>=0.4",
    "zarr>=3.0",
]

[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[tool.maturin]
# "mixed" layout: Python in python/, Rust in src/
python-source = "python"
module-name = "rust_warp._rust"
features = ["pyo3/extension-module"]

[tool.uv]
# Rebuild when Rust sources change
cache-keys = [
    { file = "pyproject.toml" },
    { file = "Cargo.toml" },
    { file = "src/**/*.rs" },
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "benchmark: marks tests as benchmarks (deselect with '-m \"not benchmark\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "stress: marks tests as stress/large-scale tests (deselect with '-m \"not stress\"')",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "PT", "RUF"]
ignore = [
    "E501",   # line length — handled by formatter
    "SIM108", # ternary operator — sometimes less readable
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]  # assert False is fine in tests

[tool.ruff.format]
# defaults are fine — matches black style

[dependency-groups]
dev = [
    "ruff>=0.15.1",
    "pre-commit>=4.0",
]
